tsup.config で awilix-types を実行できるようにutilityを追加
