+++ backend/tsup.config.ts
@@ -1,4 +1,7 @@
 import { defineConfig } from 'tsup'
+import { generateTypes } from '@workspace/type-safe-awilix/build'
+
+await generateTypes({ load: 'src/**/*.ts', outDir: 'src/generated' })
 
 export default defineConfig(() => ({
   entry: ['src/**/*.ts', '!src/**/*.test.ts'],
