+++ packages/type-safe-awilix/tsconfig.json
@@ -8,6 +8,6 @@
     "lib": ["ES2020"],
     "module": "ESNext",
     "moduleResolution": "node",
-    "types": ["node"]
+    "declaration": true
   }
 }
